+++ includes/frontend/class-header.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.0.1
+ * @version  1.0.5
  *
  * Contents:
  *
@@ -343,7 +343,7 @@ class Reykjavik_Header {
 		 * HTML Body classes
 		 *
 		 * @since    1.0.0
-		 * @version  1.0.1
+		 * @version  1.0.5
 		 *
 		 * @param  array $classes
 		 */
@@ -473,6 +473,16 @@ class Reykjavik_Header {
 							&& get_theme_mod( 'layout_page_outdent', true )
 						) {
 						$classes[] = 'page-layout-outdented';
+					}
+
+				// Enable outdented single post meta
+
+					if ( (bool) apply_filters( 'wmhook_reykjavik_header_body_classes_enable_entry_meta_outdented',
+						is_single( get_the_ID() )
+						&& ! is_active_sidebar( 'sidebar' )
+						&& ! Reykjavik_Post::is_page_builder_ready()
+					) ) {
+						$classes[] = 'entry-meta-outdented';
 					}